refactor: (carousel) separate into presentational and container components
