+++ app/components/index.jsx
@@ -1,16 +1,11 @@
 export { default as BasicProfile } from './BasicProfile';
+export { default as Carousel } from './Carousel';
 export { default as Champion } from './Champion';
 export { default as ChampionMasteries } from './ChampionMasteries';
-export { default as Filter } from './Filter';
 export { default as Footer } from './Footer';
-export { default as Home } from './Home';
 export { default as Leaderboards } from './Leaderboards';
 export { default as LiveGame } from './LiveGame';
-// export { default as LiveMatch } from './LiveMatch';
-export { default as MatchHistory } from './MatchHistory';
 export { default as Navbar } from './Navbar';
 export { default as NotFound } from './NotFound';
 export { default as Summoner } from './Summoner';
-// export { default as  } from './'
-// export { default as  } from './'
